Issue #2037213 Upgrade Qpid JMS to 2.5.0
To support Java 17.

This is Qpid JMS 2.5.0 unmodified from upstream. Previously we were
using Qpid JMS 0.59 and had patched the NettyTcpTransport class in the
qpid-jms-client JAR to support some AWIPS thin client functionality.
That patch will be applied to this version of Qpid JMS also, but for
verifiability I am applying the patch in a separate commit.

Change-Id: I8df36116aaec0227b430c28b3a14cc4de0d927e0