Issue #2038857 Upgrade Apache ActiveMQ to 6.1.6
https://nvd.nist.gov/vuln/detail/CVE-2025-27533

This release declares a dependency on Jackson 2.18.3. We are currently
shipping 2.17.2. However, there are only about 20 commits to ActiveMQ
between 6.1.3 and 6.1.6, excluding dependency upgrades. I took a look at
these and did not see anything that would break if we leave Jackson at
the current version 2.17.2. So we can avoid that upgrade for now.

Change-Id: Ic7ed3ad68110d7f71ffb675d977c0a55ee37a023